--- src/components/CharacterContainer.js
@@ -1,38 +1,0 @@
-import Character from './Character';
-
-function CharacterContainer({ characters }) {
-
-    const buildRows = () => {
-        let rows = [], cols = [];
-        characters.forEach((character, index) => {
-            cols.push(<Character key={character._id} character={character} />);
-            if ((index + 1) % 5 === 0) {
-                rows.push(
-                    <div className="character-row" key={index}>
-                        {cols}
-                    </div>
-                )
-                cols = []
-            }
-        });
-
-        // Final remaining cols 
-        if (cols.length > 0) {
-            rows.push(
-                <div className="character-row" key={characters.length}>
-                    {cols}
-                </div>
-            )
-        }
-
-        return rows;
-    }
-
-    return (
-        <div className="character-container">
-            {buildRows()}
-        </div>
-    )
-}
-
-export default CharacterContainer
