+++ src/components/character.tsx
@@ -1,6 +1,9 @@
-function Character({ character }) {
+import { DisneyCharacter } from "../disney_character"
 
-  return (
+// for our props we can reuse the DisneyCharacter interface
+// - defining an anonymous type that just has one property - a DisneyCharacter
+const Character : React.FC<{ character: DisneyCharacter}> = ( { character }) => 
+  
     <article className="character-item">
 
       <h2>{character.name}</h2>
@@ -12,7 +15,7 @@ function Character({ character }) {
       <img className="character-item__img" src={character.imageUrl} alt={character.name} />
 
     </article>
-  )
-}
+  
+
 
 export default Character
