+++ src/disney_character.ts
@@ -1,0 +1,5 @@
+export interface DisneyCharacter{
+	_id: number;
+	name: string;
+	imageUrl: string;
+}
